pyproject.toml
[project]
name = "alpaca-api"
version = "0.1.0"
description = "FastAPI wrapper for Alpaca paper trading API"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "alpaca-py",
    "pydantic-settings",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alpaca_api"]
